Exception handling - delete

+++ src/main/java/com/matheusdev/demo/services/UserService.java
@@ -1,51 +1,53 @@
 package com.matheusdev.demo.services;
 
 import java.util.List;
-
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.matheusdev.demo.entities.User;
 import com.matheusdev.demo.repositories.UserRepository;
+import com.matheusdev.demo.services.exceptions.DatabaseException;
 import com.matheusdev.demo.services.exceptions.ResourceNotFoundException;
-
 
 @Service
 public class UserService {
-	
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
 	public User findById(Long id) {
-		Optional<User> obj =  repository.findById(id);
+		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
 	
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
-	
+	    try {
+	        if (repository.existsById(id)) {
+	            repository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
+	} 
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
 		return repository.save(entity);
 	}
-
 	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-		
 	}
-
 }
